+++ internal/pkg/users/usecase/usecase.go
@@ -103,13 +103,13 @@ func (uc *UserUsecase) ValidateAndGetUser(ctx context.Context, token string) (mo
 		return models.User{}, users.ErrorUnauthorized
 	}
 
-	version, ok := claims["version"].(int)
+	version, ok := claims["version"].(float64)
 	if !ok {
 		logger.Error("invalid version claim")
 		return models.User{}, users.ErrorUnauthorized
 	}
 
-	if version != user.Version {
+	if int(version) != user.Version {
 		return models.User{}, err
 	}
